Converted process tomography to abstract base class and move circuit creation method
