+++ lightworks/tomography/process_tomography_mle.py
@@ -16,12 +16,9 @@ import warnings
 
 import numpy as np
 
-from ..sdk.circuit import Circuit
 from ..sdk.state import State
 from .process_tomography import ProcessTomography
 from .utils import (
-    INPUT_MAPPING,
-    MEASUREMENT_MAPPING,
     PAULI_MAPPING,
     RHO_MAPPING,
     unvec,
@@ -34,7 +31,25 @@ TOMO_MEASUREMENTS = ["X", "Y", "Z", "I"]
 
 class MLEProcessTomography(ProcessTomography):
     """
-    Desc
+    Runs quantum process tomography using the maximum likelihood estimation
+    method.
+
+    Args:
+
+        n_qubits (int) : The number of qubits that will be used as part of the
+            tomography.
+
+        base_circuit (Circuit) : An initial circuit which produces the required
+            output state and can be modified for performing tomography. It is
+            required that the number of circuit input modes equals 2 * the
+            number of qubits.
+
+        experiment (Callable) : A function for performing the required
+            tomography experiments. This should accept a list of circuits and a
+            list of inputs and then return a list of results to process.
+
+        experiment_args (list | None) : Optionally provide additional arguments
+            which will be passed directly to the experiment function.
     """
 
     def process(self) -> np.ndarray:
@@ -93,26 +108,6 @@ class MLEProcessTomography(ProcessTomography):
             for i, in_state in enumerate(all_inputs)
             for j, meas in enumerate(all_measurements)
         }
-
-    def _create_circuit_and_input(
-        self, input_op: str, output_op: str
-    ) -> tuple[Circuit, State]:
-        """
-        Creates the required circuit and input state to achieve a provided input
-        and output operation.
-        """
-        in_state = State([])
-        circ = Circuit(self.base_circuit.input_modes)
-        # Input operation
-        for i, op in enumerate(input_op.split(",")):
-            in_state += INPUT_MAPPING[op][0]
-            circ.add(INPUT_MAPPING[op][1], 2 * i)
-        # Add base circuit
-        circ.add(self.base_circuit)
-        # Measurement operation
-        for i, op in enumerate(output_op.split(",")):
-            circ.add(MEASUREMENT_MAPPING[op], 2 * i)
-        return circ, in_state
 
     def _generate_all_inputs(self) -> list:
         all_inputs = list(TOMO_INPUTS)
